fix some bbox types

+++ src/stats/hexbin.jl
@@ -71,7 +71,7 @@ function data_limits(hb::Hexbin)
     nw = widths(bb) .+ (ms..., 0.0f0)
     no = bb.origin .- ((ms ./ 2.0f0)..., 0.0f0)
 
-    return Rect3f(no, nw)
+    return Rect3d(no, nw)
 end
 boundingbox(p::Hexbin) = transform_bbox(p, data_limits(hb))
